Unificação de Onboarding em uma tela única + ajustes no layout e Strings
