+++ src/constants/Strings.ts
@@ -31,19 +31,28 @@ export const Strings = {
   },
 
   onboarding: {
-    logoAlt: 'Logotipo do Point Tils',
-    illoAlt: 'Ilustração da tela de boas-vindas',
-    title: 'Conecte-se a intérpretes de forma rápida e simples',
-    subtitle:
-      'Encontre intérpretes próximos, verifique disponibilidade e agende atendimentos para situações urgentes ou momentos importantes.',
-    cta: 'Encontrar intérprete agora',
-  },
-  onboardingCompany: {
-    logoAlt: 'Logotipo do Point Tils',
-    illoAlt: 'Ilustração da tela de busca de intérpretes',
-    title: 'Intérpretes certos para a sua necessidade',
-    subtitle:
-      'Pesquise e filtre intérpretes qualificados por especialidade, localização e disponibilidade, e garanta atendimento para reuniões, eventos e muito mais.',
-    cta: 'Buscar intérprete',
+    til: {
+      logoAlt: 'Logotipo do Point Tils',
+      illoAlt: 'Ilustração da tela de boas-vindas',
+      title: 'Conecte-se a intérpretes de forma rápida e simples',
+      subtitle:
+        'Encontre intérpretes próximos, verifique disponibilidade e agende atendimentos para situações urgentes ou momentos importantes.',
+      cta: 'Encontrar intérprete agora',
+    },
+    company: {
+      logoAlt: 'Logotipo do Point Tils',
+      illoAlt: 'Ilustração da tela de busca de intérpretes',
+      title: 'Intérpretes certos para a sua necessidade',
+      subtitle:
+        'Pesquise e filtre intérpretes qualificados por especialidade, localização e disponibilidade, e garanta atendimento para reuniões, eventos e muito mais.',
+      cta: 'Buscar intérprete',
+    },
+    default: {
+      logoAlt: 'Logotipo do Point Tils',
+      illoAlt: 'Ilustração padrão',
+      title: '',
+      subtitle: '',
+      cta: '',
+    },
   },
 } as const;
